refactor: 카트의 물품 삭제 로직에 인증 정보를 반영하도록 수정

+++ src/main/java/cart/dao/CartDao.java
@@ -86,8 +86,8 @@ public class CartDao implements CartRepository {
     }
 
     @Override
-    public void delete(final Long cartProductId) {
-        final String query = "DELETE FROM cart c WHERE c.id = ?";
-        jdbcTemplate.update(query, cartProductId);
+    public int delete(final Long userId, final Long cartProductId) {
+        final String query = "DELETE FROM cart c WHERE c.user_id = ? and c.id = ?";
+        return jdbcTemplate.update(query, userId, cartProductId);
     }
 }
